+++ packages/core/src/index.ts
@@ -21,10 +21,12 @@ import {
 // @ts-ignore untyped library
 import retimer from 'retimer'
 
-import { FULL_VERSION, INTERMEDIATE_HOPS, MAX_HOPS, PACKET_SIZE, VERSION } from './constants.js'
+import { FULL_VERSION, INTERMEDIATE_HOPS, MAX_HOPS, PACKET_SIZE, VERSION, MAX_PARALLEL_PINGS } from './constants.js'
 
-import NetworkPeers, { type Entry, NetworkPeersOrigin } from './network/network-peers.js'
-import Heartbeat, { NetworkHealthIndicator } from './network/heartbeat.js'
+import { Network, PeerStatus, PeerOrigin, Health, HeartbeatConfig } from '../lib/core_network.js'
+import { core_network_set_panic_hook } from '../lib/core_network.js'
+core_network_set_panic_hook()
+import Heartbeat from './network/heartbeat.js'
 
 import { findPath } from './path/index.js'
 
@@ -214,7 +216,7 @@ class Hopr extends EventEmitter {
 
   private stopPeriodicCheck: (() => void) | undefined
   private strategy: ChannelStrategyInterface
-  private networkPeers: NetworkPeers
+  private networkPeers: Network
   private heartbeat: Heartbeat
   private forward: PacketForwardInteraction
   private acknowledgements: AcknowledgementInteraction
@@ -339,16 +341,41 @@ class Hopr extends EventEmitter {
     ) => libp2pSendMessage(this.libp2pComponents, dest, protocols, msg, includeReply, opts)) as SendMessage // Typescript limitation
 
     // Attach network health measurement functionality
-    const peers: Peer[] = await this.libp2pComponents.getPeerStore().all()
-    this.networkPeers = new NetworkPeers(
-      peers.map((p) => p.id),
-      [this.id],
+
+    this.networkPeers = Network.build(
+      this.id.toString(),
       this.options.networkQualityThreshold,
-      (peer: PeerId) => {
-        this.libp2pComponents.getPeerStore().delete(peer)
-        this.publicNodesEmitter.emit('removePublicNode', peer)
-      }
+      (peer: string) => {
+        let p = peerIdFromString(peer)
+        this.libp2pComponents.getPeerStore().delete(p)
+        this.publicNodesEmitter.emit('removePublicNode', p)
+      },
+      (oldHealth: Health, newHealth: Health) => this.emit('hopr:network-health-changed', oldHealth, newHealth),
+      (peerId: string) => {
+        let peer = peerIdFromString(peerId)
+        if (this.knownPublicNodesCache.has(peerId)) return true
+
+        // If we have a direct connection to this peer ID, declare it a public node
+        if (
+          libp2p.connectionManager
+            .getConnections(peer)
+            .flatMap((c) => c.tags ?? [])
+            .includes(PeerConnectionType.DIRECT)
+        ) {
+          this.knownPublicNodesCache.add(peerId)
+          return true
+        }
+
+        return false
+      },
+      (peer: string) => this.closeConnectionsTo(peerIdFromString(peer))
     )
+
+    // initialize with all the peers identified in the peer store
+    const peers: Peer[] = await this.libp2pComponents.getPeerStore().all()
+    for (const peer in peers.map((p) => p.id)) {
+      this.networkPeers.register(peer.toString(), PeerOrigin.Initialization)
+    }
 
     // react when network registry is enabled / disabled
     connector.indexer.on('network-registry-status-changed', async (enabled: boolean) => {
@@ -357,7 +384,7 @@ class Hopr extends EventEmitter {
       if (enabled) {
         for (const connection of this.libp2pComponents.getConnectionManager().getConnections()) {
           if (!(await this.isAllowedAccessToNetwork(connection.remotePeer))) {
-            this.networkPeers.unregister(connection.remotePeer)
+            this.networkPeers.unregister(connection.remotePeer.toString())
             try {
               await connection.close()
             } catch (err) {
@@ -376,7 +403,7 @@ class Hopr extends EventEmitter {
         // otherwise there is nothing to do
         if (!eligible) {
           for (const node of nodes) {
-            this.networkPeers.unregister(node.toPeerId())
+            this.networkPeers.unregister(node.toPeerId().toString())
 
             for (const conn of this.libp2pComponents.getConnectionManager().getConnections(node.toPeerId())) {
               try {
@@ -392,36 +419,23 @@ class Hopr extends EventEmitter {
 
     peers.forEach((peer) => log(`peer store: loaded peer ${peer.id.toString()}`))
 
+    let heartbeat_config = HeartbeatConfig.build(
+      MAX_PARALLEL_PINGS,
+      this.options?.heartbeatVariance,
+      this.options?.heartbeatInterval,
+      BigInt(this.options?.heartbeatThreshold)
+    )
+
     this.heartbeat = new Heartbeat(
-      this.id,
       this.networkPeers,
       this.libp2pComponents,
       sendMessage,
-      this.closeConnectionsTo.bind(this),
-      (oldHealthValue: NetworkHealthIndicator, newNetworkHealth: NetworkHealthIndicator) =>
-        this.emit('hopr:network-health-changed', oldHealthValue, newNetworkHealth),
-      (peerId: PeerId) => {
-        if (this.knownPublicNodesCache.has(peerId.toString())) return true
-
-        // If we have a direct connection to this peer ID, declare it a public node
-        if (
-          libp2p.connectionManager
-            .getConnections(peerId)
-            .flatMap((c) => c.tags ?? [])
-            .includes(PeerConnectionType.DIRECT)
-        ) {
-          this.knownPublicNodesCache.add(peerId.toString())
-          return true
-        }
-
-        return false
-      },
       this.environment.id,
-      this.options
+      heartbeat_config
     )
 
     this.libp2pComponents.getConnectionManager().addEventListener('peer:connect', (event: CustomEvent<Connection>) => {
-      this.networkPeers.register(event.detail.remotePeer, NetworkPeersOrigin.INCOMING_CONNECTION)
+      this.networkPeers.register(event.detail.remotePeer.toString(), PeerOrigin.IncomingConnection)
     })
 
     this.acknowledgements = new AcknowledgementInteraction(
@@ -501,7 +515,6 @@ class Hopr extends EventEmitter {
       log(`No multiaddrs has been registered.`)
     }
     await this.maybeLogProfilingToGCloud()
-    this.heartbeat.recalculateNetworkHealth()
   }
 
   private async maybeLogProfilingToGCloud() {
@@ -610,9 +623,7 @@ class Hopr extends EventEmitter {
       }
     }
 
-    // Mark the corresponding entry as public & recalculate network health indicator
     this.knownPublicNodesCache.add(peer.id.toString())
-    this.heartbeat.recalculateNetworkHealth()
   }
 
   private async strategyOpenChannel(status: OutgoingChannelStatus) {
@@ -621,7 +632,7 @@ class Hopr extends EventEmitter {
       const stake = new BN(status.stake_str)
 
       if (await this.isAllowedAccessToNetwork(destination)) {
-        this.networkPeers.register(destination, NetworkPeersOrigin.STRATEGY_NEW_CHANNEL)
+        this.networkPeers.register(destination.toString(), PeerOrigin.StrategyNewChannel)
 
         const hash = await this.openChannel(destination, stake)
         verbose('- opened channel', destination, hash)
@@ -695,7 +706,7 @@ class Hopr extends EventEmitter {
       await Promise.all(
         outgoingChannels.map(async (channel) => {
           if (await this.isAllowedAccessToNetwork(channel.destination.toPeerId())) {
-            this.networkPeers.register(channel.destination.toPeerId(), NetworkPeersOrigin.STRATEGY_EXISTING_CHANNEL)
+            this.networkPeers.register(channel.destination.toPeerId().toString(), PeerOrigin.StrategyExistingChannel)
           } else {
             error(`Protocol error: Strategy is monitoring non-registered peer ${channel.destination.toString()}`)
           }
@@ -705,10 +716,7 @@ class Hopr extends EventEmitter {
       // Perform the strategy tick
       tickResult = this.strategy.tick(
         (await this.getBalance()).toBN(),
-        this.networkPeers
-          .all()
-          .map((p) => p.toString())
-          .values(),
+        this.networkPeers.all().values(),
         outgoingChannels.map((c) => {
           return {
             peer_id: c.destination.toPeerId().toString(),
@@ -716,7 +724,7 @@ class Hopr extends EventEmitter {
             status: c.status
           }
         }),
-        (peer_id_str: string) => this.networkPeers.qualityOf(peerIdFromString(peer_id_str))
+        (peer_id_str: string) => this.networkPeers.quality_of(peer_id_str)
       )
       metric_strategyTicks.increment()
       metric_strategyMaxChannels.set(tickResult.max_auto_channels)
@@ -752,7 +760,7 @@ class Hopr extends EventEmitter {
    * Recalculates and retrieves the current connectivity health indicator.
    */
   public getConnectivityHealth() {
-    return this.heartbeat.recalculateNetworkHealth()
+    return this.networkPeers.health()
   }
 
   /**
@@ -949,16 +957,17 @@ class Hopr extends EventEmitter {
     if (!(await this.isAllowedAccessToNetwork(destination))) {
       throw Error(`Connection to node is not allowed`)
     }
-    // Propagate any errors thrown upwards
-    let pingResult = await this.heartbeat.pingNode(destination)
 
-    if (pingResult.lastSeen >= 0) {
-      if (this.networkPeers.has(destination)) {
-        this.networkPeers.updateRecord(pingResult)
-      } else {
-        this.networkPeers.register(destination, NetworkPeersOrigin.MANUAL_PING)
-      }
-      return { latency: pingResult.lastSeen - start }
+    let dest = destination.toString()
+    if (!this.networkPeers.contains(dest)) {
+      this.networkPeers.register(dest, PeerOrigin.ManualPing)
+    }
+
+    await this.heartbeat.pingNode(destination)
+
+    let peer_info = this.networkPeers.get_peer_info(destination.toString())
+    if (peer_info.last_seen >= 0) {
+      return { latency: Number(peer_info.last_seen) - start }
     } else {
       return { info: 'failure', latency: -1 }
     }
@@ -972,10 +981,10 @@ class Hopr extends EventEmitter {
       return []
     }
 
-    const entries = this.networkPeers.getAllEntries()
+    const entries = this.networkPeers.all()
     return (function* () {
       for (const entry of entries) {
-        yield entry.id
+        yield peerIdFromString(entry)
       }
     })()
   }
@@ -992,8 +1001,8 @@ class Hopr extends EventEmitter {
    * @param peerId of the node we want to get the connection info for
    * @returns various information about the connection
    */
-  public getConnectionInfo(peerId: PeerId): Entry {
-    return this.networkPeers.getConnectionInfo(peerId)
+  public getConnectionInfo(peerId: PeerId): PeerStatus {
+    return this.networkPeers.get_peer_info(peerId.toString())
   }
 
   /**
@@ -1026,7 +1035,7 @@ class Hopr extends EventEmitter {
     if (!this.networkPeers) {
       return 'Node has not started yet'
     }
-    const connected = this.networkPeers.debugLog()
+    const connected = this.networkPeers.debug_output()
 
     let announced: string[] = []
     for await (const announcement of HoprCoreEthereum.getInstance().indexer.getAddressesAnnouncedOnChain()) {
@@ -1446,7 +1455,7 @@ class Hopr extends EventEmitter {
       PublicKey.fromPeerId(this.getId()),
       destination,
       hops,
-      (p: PublicKey) => this.networkPeers.qualityOf(p.toPeerId()),
+      (p: PublicKey) => this.networkPeers.quality_of(p.toPeerId().toString()),
       HoprCoreEthereum.getInstance().getOpenChannelsFrom.bind(HoprCoreEthereum.getInstance())
     )
   }
@@ -1519,8 +1528,9 @@ export {
   isStrategy,
   SaneDefaults,
   findPath,
-  NetworkHealthIndicator,
-  NetworkPeersOrigin,
+  PeerOrigin,
+  PeerStatus,
+  Health,
   type ChannelStrategyInterface
 }
 export { resolveEnvironment, supportedEnvironments, type ResolvedEnvironment } from './environment.js'